Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,8 +2,8 @@
 #include<stdio.h>
 
 int main(){
-    int radius;
-    scanf("%.2f",&radius);
+    float  radius;
+    scanf("%f",&radius);
     printf("Area: %.2f", radius * radius * 3.14);
     return 0;
 }
